configure directory
