+++ simulation/behavior_tree_plugin/CMakeLists.txt
@@ -14,13 +14,13 @@ find_package(pluginlib REQUIRED)
 
 add_library(behavior_tree_plugin SHARED
   src/action_node.cpp
-  src/logger.cpp
   src/pedestrian/behavior_tree.cpp
   src/pedestrian/follow_lane_action.cpp
   src/pedestrian/pedestrian_action_node.cpp
   src/pedestrian/walk_straight_action.cpp
-  src/reset_request.cpp
-  src/transition_event.cpp
+  src/transition_events/logger.cpp
+  src/transition_events/reset_request.cpp
+  src/transition_events/transition_event.cpp
   src/vehicle/behavior_tree.cpp
   src/vehicle/follow_lane_sequence/follow_front_entity_action.cpp
   src/vehicle/follow_lane_sequence/follow_lane_action.cpp
