+++ simulation/behavior_tree_plugin/include/behavior_tree_plugin/transition_events/reset_request.hpp
@@ -12,28 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef BEHAVIOR_TREE_PLUGIN__LOGGER_HPP_
-#define BEHAVIOR_TREE_PLUGIN__LOGGER_HPP_
+#ifndef BEHAVIOR_TREE_PLUGIN__TRANSITION_EVENTS__RESET_REQUEST_HPP_
+#define BEHAVIOR_TREE_PLUGIN__TRANSITION_EVENTS__RESET_REQUEST_HPP_
 
-#include <behavior_tree_plugin/transition_event.hpp>
+#include <behavior_tree_plugin/transition_events/transition_event.hpp>
 #include <memory>
 #include <rclcpp/rclcpp.hpp>
 
 namespace behavior_tree_plugin
 {
-class Logger : public TransitionEvent
+class ResetRequest : public TransitionEvent
 {
 public:
-  Logger(const std::shared_ptr<BT::Tree> & tree_ptr, const rclcpp::Logger & logger);
+  ResetRequest(
+    const std::shared_ptr<BT::Tree> & tree_ptr, std::function<std::string()> get_request_function,
+    std::function<void(std::string)> set_request_function);
   const std::string getCurrentAction() const;
 
 private:
   void callback(
     BT::Duration timestamp, const BT::TreeNode & node, BT::NodeStatus prev_status,
     BT::NodeStatus status) override;
-  rclcpp::Logger ros_logger_;
-  std::string current_action_;
+  std::shared_ptr<BT::Tree> tree_ptr_;
+  std::function<std::string()> get_request_function_;
+  std::function<void(std::string)> set_request_function_;
 };
 }  // namespace behavior_tree_plugin
 
-#endif  // BEHAVIOR_TREE_PLUGIN__LOGGER_HPP_
+#endif  // BEHAVIOR_TREE_PLUGIN__TRANSITION_EVENTS_RESET__REQUEST_HPP_
